Add Auth0 login link and minimal account details
I have a bad habit of not keeping up with the documentation as I go. I'm
adding minimal notes and links as I progress in an effort to break this
habit.

This information **should** be sufficient to get someone started on the
login once it is fully fuctional.  At the time of writing, it is missing
the necessary roles to complete the login.  I am going to move the
terraform that handles these from `kubernetes-investigations` to here
next.