Set up circleci to validate terraform
This is the simplest validation possible.  I decided against linting at
this point as, despite the TF coming back as invalid, `wata727/tflint`
was not rasing any exceptions.  I believe linting will be valuable
moving forward.  However, what I want right now is to catch the obvious
mistakes.

I've used a shell script as the `hashicorp/terraform` alpine container
does not come with `make`, which would have been my first choice.  I
organised the tests in this way so as to allow for different TF
subdirectories and tools written in other languages.