Initial setup (#1)
* Change title to use Titlecase

A minor stylistic issue.

* Add Auth0 login link and minimal account details

I have a bad habit of not keeping up with the documentation as I go. I'm
adding minimal notes and links as I progress in an effort to break this
habit.

This information **should** be sufficient to get someone started on the
login once it is fully fuctional.  At the time of writing, it is missing
the necessary roles to complete the login.  I am going to move the
terraform that handles these from `kubernetes-investigations` to here
next.

* Ensure saml xml metadata gets encrypted

It contains all the secrets for this setup and so should never be
stored unencrypted.

* Add the actual saml xml metadata file

Now that the git-crypt config is in place.  I've decided **not** to
template this as it can be download complete from Auth0 and it seems
like it should be the source of truth.

* Add the simplest .tf possible

This was done in a module in
`ministryofjustice/kubernetes-investigations`.  It seemed simipler to
break it out into a very simple `main.tf`.  That is also essentially
what Kerin recommended when he was prototyping that setup.

* Set up circleci to validate terraform

This is the simplest validation possible.  I decided against linting at
this point as, despite the TF coming back as invalid, `wata727/tflint`
was not rasing any exceptions.  I believe linting will be valuable
moving forward.  However, what I want right now is to catch the obvious
mistakes.

I've used a shell script as the `hashicorp/terraform` alpine container
does not come with `make`, which would have been my first choice.  I
organised the tests in this way so as to allow for different TF
subdirectories and tools written in other languages.