Add an outline README covering GitHub/Auth0 rules
It's a pretty complex process getting them wired in. As a result it is
easily forgotten and steps missed.  This *should* be the minimum
configutation between Auth0 and GitHub that will work-I say 'should'
because I can easily imagine things changing between the time of writing
and the time someone has to review and use the steps.