Auth0 rules (#2)
* Port Auth0 JS rules from kubernetes-investigations

Done so everything related is in the same repo and so it can be adapted
for the narrower use-case of this repo.

* Add a terrform format check to ./test.sh

* Remove the OIDC rules

These are conceptually simpler than the SAML rules, but Auth0 seems
better set up for dealing with SAML.

* Change saml team mapping to only use github-webops

This is the only team->role mapping we need right now.  I have also
changed the hardcoded ARNs to point at the correct info on the correct
account.

* Add simple github to oauth mapping

This will give anyone in the webops team admin permissions on the new
AWS account.  It depends on two earlier rules to ensure they are part of
the MoJ and in the correct team.

* Add whitelisting for webops team

This is the quick and dirty way of ensuring only webops has admin
access to the new AWS account for now.  It should also be able to
authorise lower-level roles for other departmental users.  This will be
for a later refactor, though.

* Add an outline README covering GitHub/Auth0 rules

It's a pretty complex process getting them wired in. As a result it is
easily forgotten and steps missed.  This *should* be the minimum
configutation between Auth0 and GitHub that will work-I say 'should'
because I can easily imagine things changing between the time of writing
and the time someone has to review and use the steps.